HW01 and Optional: changes on review

+++ src/main/java/ru/javawebinar/topjava/web/MealServlet.java
@@ -2,16 +2,18 @@ package ru.javawebinar.topjava.web;
 
 import org.slf4j.Logger;
 import ru.javawebinar.topjava.model.Meal;
-import ru.javawebinar.topjava.model.RuntimeMealDao;
+import ru.javawebinar.topjava.dao.RuntimeMealDao;
 import ru.javawebinar.topjava.model.MealTo;
 import ru.javawebinar.topjava.util.MealsUtil;
 
+import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.List;
@@ -19,42 +21,16 @@ import java.util.List;
 import static org.slf4j.LoggerFactory.getLogger;
 
 public class MealServlet extends HttpServlet {
-    private static final String ACTION_NAME="action";
+    private static final String ACTION_NAME = "action";
+    private static final String MEALS_LIST = "/topjava/meals";
     private static final Logger log = getLogger(MealServlet.class);
     private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm");
-    RuntimeMealDao dao = new RuntimeMealDao();
+    private RuntimeMealDao dao;
 
-    private Integer getIntegerParam(HttpServletRequest request, String paramName) {
-        String paramStr = request.getParameter(paramName);
-        Integer result = null;
-        if (paramStr != null && !paramStr.isEmpty())
-            try {
-                result = Integer.valueOf(paramStr);
-            } catch (NumberFormatException e) {
-                log.warn("'{}' invalid parameter format", paramName);
-            }
-        return result;
-    }
-
-    private LocalDateTime getDateTimeParam(HttpServletRequest request, String paramName) {
-        String paramStr = request.getParameter(paramName);
-        LocalDateTime result = null;
-        if (paramStr != null && !paramStr.isEmpty()) {
-            try {
-                result = LocalDateTime.parse(paramStr, formatter);
-            } catch (DateTimeParseException e) {
-                log.warn("'{}' invalid datetime parameter format", paramName);
-            }
-        }
-        return result;
-    }
-
-    private void getMealsTo(HttpServletRequest request) {
-        Integer caloriesLimit = getIntegerParam(request, "limit");
-        List<MealTo> meals = MealsUtil.filter(dao.getAll(),  //MealsUtil.getMeals(),
-                null, null,
-                caloriesLimit != null ? caloriesLimit.intValue() : 2000);
-        request.setAttribute("list", meals);
+    @Override
+    public void init(ServletConfig config) throws ServletException{
+        super.init(config);
+        dao = new RuntimeMealDao();
     }
 
     @Override
@@ -76,10 +52,12 @@ public class MealServlet extends HttpServlet {
                 break;
             case "delete":
                 Integer deletedId = getIntegerParam(request, "id");
-                log.debug("Delete meal {}", deletedId);
-                dao.delete(deletedId);
+                if (deletedId != null) {
+                    log.debug("Delete meal {}", deletedId);
+                    dao.delete(deletedId);
+                }
                 getMealsTo(request);
-                request.getRequestDispatcher("/meals.jsp").forward(request, response);
+                response.sendRedirect(MEALS_LIST);
                 break;
             case "list":
             default:
@@ -94,20 +72,47 @@ public class MealServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         log.debug("Post");
         request.setCharacterEncoding("UTF-8");
-        String action = request.getParameter(ACTION_NAME);
-        if ("submit".equals(action)) {
-            Integer calories = getIntegerParam(request, "calories");
-            Integer id = getIntegerParam(request, "id");
-            LocalDateTime dateTime = getDateTimeParam(request, "datetime");
-            log.debug("Update meal {}", id);
-            Meal newMeal = new Meal(id, dateTime, request.getParameter("description"), calories != null ? calories.intValue() : 0);
-            if (id == null)
-                dao.add(newMeal);
-            else
-                dao.update(newMeal);
-        }
+        Integer calories = getIntegerParam(request, "calories");
+        Integer id = getIntegerParam(request, "id");
+        LocalDateTime dateTime = getDateTimeParam(request, "datetime");
+        log.debug("Update meal {}", id);
+        Meal newMeal = new Meal(id, dateTime, request.getParameter("description"), calories != null ? calories.intValue() : 0);
+        if (id == null)
+            dao.add(newMeal);
+        else
+            dao.update(newMeal);
         log.debug("Post meals list");
         getMealsTo(request);
-        request.getRequestDispatcher("/meals.jsp").forward(request, response);
+        response.sendRedirect(MEALS_LIST);
+    }
+
+    private Integer getIntegerParam(HttpServletRequest request, String paramName) {
+        String paramStr = request.getParameter(paramName);
+        if (paramStr != null && !paramStr.isEmpty())
+            try {
+                return Integer.valueOf(paramStr);
+            } catch (NumberFormatException e) {
+                log.warn("'{}' invalid parameter format", paramName);
+            }
+        return null;
+    }
+
+    private LocalDateTime getDateTimeParam(HttpServletRequest request, String paramName) {
+        String paramStr = request.getParameter(paramName);
+        if (paramStr != null && !paramStr.isEmpty()) {
+            try {
+                return LocalDateTime.parse(paramStr, formatter);
+            } catch (DateTimeParseException e) {
+                log.warn("'{}' invalid datetime parameter format", paramName);
+            }
+        }
+        return null;
+    }
+
+    private void getMealsTo(HttpServletRequest request) {
+        List<MealTo> meals = MealsUtil.filteredByStreams(dao.getAll(),  //MealsUtil.getMeals(),
+                LocalTime.MIN, LocalTime.MAX,
+                2000);
+        request.setAttribute("list", meals);
     }
 }
